requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.13.0
google-cloud-firestore==2.14.0
google-cloud-core==2.4.1
google-auth==2.23.4
gunicorn==21.2.0
python-dotenv
structlog
google-cloud-bigquery-storage